Delete explorer_common/resource_utils.js, move get_resources_for_subject from it to resource_explorer_common.js
