--- client/src/explorer_common/resource_utils.js
@@ -1,12 +1,0 @@
-import { results_resource_fragment } from '../panels/panel_declarations/results/results_common.js';
-
-
-export function get_resources_for_subject(subject, doc){
-  const resources = results_resource_fragment(subject, doc);
-
-  if(resources.spending || resources.ftes){
-    return resources;
-  } else {
-    return null;
-  }
-}
